fix some warnings
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@47375 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/httpd/modules/upgrade_rb500.c
@@ -74,7 +74,7 @@ do_upgrade_cgi(unsigned char method, struct mime_handler *handler, char *url, we
 
 static int
 // sys_upgrade(char *url, FILE *stream, int *total)
-sys_upgrade(char *url, webs_t stream, int *total, int type)	// jimmy,
+sys_upgrade(char *url, webs_t stream, size_t *total, int type)	// jimmy,
 								// https,
 								// 8/6/2003
 {
